Created the initial project solution
Don't put the chatbot class in the Main.cpp file.  We are only building
a console application for the prototype presentation.  But after the
prototype we will be moving several sections of the code over into a
windows forms application, or some other type of capable GUI environment
that we can use to develop the game.